Shard the dataset for small sample buffer sizes
When the sample buffer size is small compared to the training data
size and the evaluation frequency is high, the training will
consistently miss a portion of the dataset.

This change randomly selects a shard of the dataset to ensure that
the full dataset can be seen during the training.