Fix model configuration loading
The previous loading approach was misleading as the path was simply
transformed into a module name (replace "/" with ".") and had the same
limitations as Python imports (e.g. no relative imports).

It is simpler and more usable to simply extend the Python search path
to the directory that contains the configuration file and import the
filename directly.

This reverts aa6ce9d9.