Make injection of special sequence tokens (<s>, </s>) configurable (#535)
This is especially needed for language models where these special
tokens may or may not be wanted depending on the task requirement.

For example, to allow nonconditional and unbounded generation of text,
one would enable <s> but disable </s>. However, to generate a fixed
sequence from a prefix/context, one would disable <s> but enable </s>.

This change also enables this configuration for seq2seq models (source
and target). There could be reasons to want </s> at the end of the
source sequences.