Avoid deepcopying the model instance
Calling deepcopy on a RNN model broke in TensorFlow 2.2. In general,
it seems this is something we should avoid. In this change, we pass a
model function to the Runner class so that it can build new model
instances on demand.

This is a partial workaround for the issue described in
https://github.com/OpenNMT/OpenNMT-tf/issues/672 as it only fixes the
error for users training with onmt-main but not for users of the
Runner class.