Raise errors when starting the training but end conditions are met (#716)
Prior to this change, a warning was logged if the training end
conditions are met before starting the training loop. However, it is a
bit unexpected for the user that the training terminates without
errors but no iterations are executed.

The error messages have been updated with the actions required to make
the training continue.