Cleanup asset prefix generation and usage (#730)
`asset_prefix` is used to differentiate resources used by parallel
inputters.

The goal of this change is to allow inputters to read the data
configuration as if there was no prefix. Ideally they should not know
that they are used with other inputters.

Now the `asset_prefix` attribute is correctly set for all parallel
inputters and the data configuration is wrapped to allow making
relative lookups (e.g. "vocabulary" will resolve to
"source_vocabulary").