Remove direct dependency to TensorFlow
There are multiple official and community TensorFlow packages out
there: tensorflow, tensorflow-cpu, tensorflow-gpu, tf-nightly,
tensorflow-rocm, intel-tensorflow, tensorflow-macos, etc.

The main "tensorflow" package works in most cases but ultimately we
should allow users to install an alternative package.

The TensorFlow version will be checked on import and a warning will be
raised if the version is not officially supported.